Refactor app module packages
