+++ app/src/debug/kotlin/com/gdavidpb/tuindice/data/DebugReportingDataSource.kt
@@ -1,0 +1,22 @@
+package com.gdavidpb.tuindice.data
+
+import android.util.Log
+import com.gdavidpb.tuindice.base.domain.repository.ReportingRepository
+
+class DebugReportingDataSource : ReportingRepository {
+    override fun setIdentifier(identifier: String) {
+        Log.d("setIdentifier", identifier)
+    }
+
+    override fun logException(throwable: Throwable) {
+        throwable.printStackTrace()
+    }
+
+    override fun logMessage(message: String) {
+        Log.d("log", message)
+    }
+
+    override fun <T : Any> setCustomKey(key: String, value: T) {
+        Log.d("setCustomKey", "$key = $value (${value::class.java.name})")
+    }
+}
