+++ app/src/debug/kotlin/com/gdavidpb/tuindice/di/AppMockModule.kt
@@ -19,12 +19,12 @@ import com.gdavidpb.tuindice.data.AuthMockDataSource
 import com.gdavidpb.tuindice.data.DebugKoinDataSource
 import com.gdavidpb.tuindice.data.MessagingMockDataSource
 import com.gdavidpb.tuindice.data.RemoteConfigMockDataSource
-import com.gdavidpb.tuindice.data.android.AndroidApplicationDataSource
-import com.gdavidpb.tuindice.data.crashlytics.DebugReportingDataSource
-import com.gdavidpb.tuindice.data.google.GooglePlayServicesDataSource
-import com.gdavidpb.tuindice.data.network.AndroidNetworkDataSource
-import com.gdavidpb.tuindice.data.retrofit.AuthorizationInterceptor
-import com.gdavidpb.tuindice.data.settings.PreferencesDataSource
+import com.gdavidpb.tuindice.data.source.application.AndroidApplicationDataSource
+import com.gdavidpb.tuindice.data.DebugReportingDataSource
+import com.gdavidpb.tuindice.data.source.mobile.GooglePlayServicesDataSource
+import com.gdavidpb.tuindice.data.source.network.AndroidNetworkDataSource
+import com.gdavidpb.tuindice.utils.AuthorizationInterceptor
+import com.gdavidpb.tuindice.data.source.settings.PreferencesDataSource
 import com.gdavidpb.tuindice.domain.usecase.GetUpdateInfoUseCase
 import com.gdavidpb.tuindice.domain.usecase.RequestReviewUseCase
 import com.gdavidpb.tuindice.domain.usecase.SetLastScreenUseCase
