--- app/src/main/kotlin/com/gdavidpb/tuindice/data/api/MessagingApi.kt
@@ -1,12 +1,0 @@
-package com.gdavidpb.tuindice.data.api
-
-import retrofit2.Response
-import retrofit2.http.POST
-import retrofit2.http.Query
-
-interface MessagingApi {
-	@POST("messaging")
-	suspend fun enroll(
-		@Query("messagingToken") messagingToken: String
-	): Response<Unit>
-}
