--- app/src/main/kotlin/com/gdavidpb/tuindice/data/crashlytics/CrashlyticsReportingDataSource.kt
@@ -1,32 +1,0 @@
-package com.gdavidpb.tuindice.data.crashlytics
-
-import com.gdavidpb.tuindice.base.domain.repository.ReportingRepository
-import com.google.firebase.crashlytics.FirebaseCrashlytics
-
-class CrashlyticsReportingDataSource(
-    private val crashlytics: FirebaseCrashlytics
-) : ReportingRepository {
-    override fun setIdentifier(identifier: String) {
-        crashlytics.setUserId(identifier)
-    }
-
-    override fun logException(throwable: Throwable) {
-        crashlytics.recordException(throwable)
-    }
-
-    override fun logMessage(message: String) {
-        crashlytics.log(message)
-    }
-
-    override fun <T : Any> setCustomKey(key: String, value: T) {
-        when (value) {
-            is Int -> crashlytics.setCustomKey(key, value)
-            is Long -> crashlytics.setCustomKey(key, value)
-            is Float -> crashlytics.setCustomKey(key, value)
-            is Double -> crashlytics.setCustomKey(key, value)
-            is String -> crashlytics.setCustomKey(key, value)
-            is Boolean -> crashlytics.setCustomKey(key, value)
-            else -> throw IllegalArgumentException("Unsupported value '$value' of type '${value::class.java.name}'")
-        }
-    }
-}
