--- app/src/main/kotlin/com/gdavidpb/tuindice/data/messaging/source/LocalDataSource.kt
@@ -1,10 +1,0 @@
-package com.gdavidpb.tuindice.data.messaging.source
-
-interface LocalDataSource {
-	suspend fun isEnrolled(): Boolean
-	suspend fun markAsEnrolled()
-
-	suspend fun getSubscribedTopics(): List<String>
-	suspend fun saveSubscriptionTopic(topic: String)
-	suspend fun isSubscribedToTopic(topic: String): Boolean
-}
