--- app/src/main/kotlin/com/gdavidpb/tuindice/data/messaging/source/ProviderDataSource.kt
@@ -1,7 +1,0 @@
-package com.gdavidpb.tuindice.data.messaging.source
-
-interface ProviderDataSource {
-	suspend fun getToken(): String?
-	suspend fun subscribeToTopic(topic: String)
-	suspend fun unsubscribeFromTopic(topic: String)
-}
