--- app/src/main/kotlin/com/gdavidpb/tuindice/data/messaging/source/RemoteDataSource.kt
@@ -1,5 +1,0 @@
-package com.gdavidpb.tuindice.data.messaging.source
-
-interface RemoteDataSource {
-	suspend fun enroll(messagingToken: String)
-}
