+++ app/src/main/kotlin/com/gdavidpb/tuindice/data/repository/messaging/LocalDataSource.kt
@@ -1,0 +1,10 @@
+package com.gdavidpb.tuindice.data.repository.messaging
+
+interface LocalDataSource {
+	suspend fun isEnrolled(): Boolean
+	suspend fun markAsEnrolled()
+
+	suspend fun getSubscribedTopics(): List<String>
+	suspend fun saveSubscriptionTopic(topic: String)
+	suspend fun isSubscribedToTopic(topic: String): Boolean
+}
