+++ app/src/main/kotlin/com/gdavidpb/tuindice/data/repository/messaging/ProviderDataSource.kt
@@ -1,0 +1,7 @@
+package com.gdavidpb.tuindice.data.repository.messaging
+
+interface ProviderDataSource {
+	suspend fun getToken(): String?
+	suspend fun subscribeToTopic(topic: String)
+	suspend fun unsubscribeFromTopic(topic: String)
+}
