+++ app/src/main/kotlin/com/gdavidpb/tuindice/data/repository/messaging/RemoteDataSource.kt
@@ -1,0 +1,5 @@
+package com.gdavidpb.tuindice.data.repository.messaging
+
+interface RemoteDataSource {
+	suspend fun enroll(messagingToken: String)
+}
