+++ app/src/main/kotlin/com/gdavidpb/tuindice/data/source/auth/mapper/FirebaseUser.kt
@@ -1,0 +1,9 @@
+package com.gdavidpb.tuindice.data.source.auth.mapper
+
+import com.gdavidpb.tuindice.base.domain.model.Auth
+import com.google.firebase.auth.FirebaseUser
+
+fun FirebaseUser.toAuth() = Auth(
+	uid = uid,
+	email = email ?: error("email")
+)
