+++ app/src/main/kotlin/com/gdavidpb/tuindice/data/source/settings/PreferencesDataSource.kt
@@ -1,0 +1,39 @@
+package com.gdavidpb.tuindice.data.source.settings
+
+import android.content.SharedPreferences
+import androidx.core.content.edit
+import com.gdavidpb.tuindice.base.domain.repository.SettingsRepository
+import com.gdavidpb.tuindice.base.presentation.navigation.Destination
+import com.gdavidpb.tuindice.base.utils.PreferencesKeys
+
+class PreferencesDataSource(
+	private val sharedPreferences: SharedPreferences
+) : SettingsRepository {
+	override fun getLastScreen(): String {
+		return sharedPreferences
+			.getString(PreferencesKeys.LAST_SCREEN, null)
+			?: Destination.Summary.route
+	}
+
+	override fun setLastScreen(route: String) {
+		sharedPreferences.edit {
+			putString(PreferencesKeys.LAST_SCREEN, route)
+		}
+	}
+
+	override fun isReviewSuggested(value: Int): Boolean {
+		val counter = sharedPreferences.getInt(PreferencesKeys.SYNCS_COUNTER, 0) + 1
+
+		sharedPreferences.edit {
+			putInt(PreferencesKeys.SYNCS_COUNTER, counter)
+		}
+
+		return counter == value
+	}
+
+	override fun clear() {
+		sharedPreferences.edit {
+			clear()
+		}
+	}
+}
