+++ app/src/main/kotlin/com/gdavidpb/tuindice/di/AppModule.kt
@@ -17,23 +17,23 @@ import com.gdavidpb.tuindice.base.domain.repository.SettingsRepository
 import com.gdavidpb.tuindice.base.utils.ResourceResolver
 import com.gdavidpb.tuindice.base.utils.extension.create
 import com.gdavidpb.tuindice.base.utils.extension.sharedPreferences
-import com.gdavidpb.tuindice.data.android.AndroidApplicationDataSource
-import com.gdavidpb.tuindice.data.api.ApiDataSource
-import com.gdavidpb.tuindice.data.api.MessagingApi
-import com.gdavidpb.tuindice.data.config.RemoteConfigDataSource
-import com.gdavidpb.tuindice.data.crashlytics.CrashlyticsReportingDataSource
-import com.gdavidpb.tuindice.data.firebase.FirebaseAuthDataSource
-import com.gdavidpb.tuindice.data.firebase.FirebaseMessagingDataSource
-import com.gdavidpb.tuindice.data.google.GooglePlayServicesDataSource
-import com.gdavidpb.tuindice.data.koin.ReleaseKoinDataSource
-import com.gdavidpb.tuindice.data.messaging.MessagingDataRepository
-import com.gdavidpb.tuindice.data.messaging.source.LocalDataSource
-import com.gdavidpb.tuindice.data.messaging.source.ProviderDataSource
-import com.gdavidpb.tuindice.data.messaging.source.RemoteDataSource
-import com.gdavidpb.tuindice.data.network.AndroidNetworkDataSource
-import com.gdavidpb.tuindice.data.preferences.MessagingPreferencesDataSource
-import com.gdavidpb.tuindice.data.retrofit.AuthorizationInterceptor
-import com.gdavidpb.tuindice.data.settings.PreferencesDataSource
+import com.gdavidpb.tuindice.data.source.application.AndroidApplicationDataSource
+import com.gdavidpb.tuindice.data.repository.messaging.source.ApiDataSource
+import com.gdavidpb.tuindice.data.repository.messaging.MessagingApi
+import com.gdavidpb.tuindice.data.source.config.RemoteConfigDataSource
+import com.gdavidpb.tuindice.data.source.reporting.CrashlyticsReportingDataSource
+import com.gdavidpb.tuindice.data.source.auth.FirebaseAuthDataSource
+import com.gdavidpb.tuindice.data.repository.messaging.source.FirebaseMessagingDataSource
+import com.gdavidpb.tuindice.data.source.mobile.GooglePlayServicesDataSource
+import com.gdavidpb.tuindice.data.source.di.ReleaseKoinDataSource
+import com.gdavidpb.tuindice.data.repository.messaging.MessagingDataRepository
+import com.gdavidpb.tuindice.data.repository.messaging.LocalDataSource
+import com.gdavidpb.tuindice.data.repository.messaging.ProviderDataSource
+import com.gdavidpb.tuindice.data.repository.messaging.RemoteDataSource
+import com.gdavidpb.tuindice.data.source.network.AndroidNetworkDataSource
+import com.gdavidpb.tuindice.data.repository.messaging.source.MessagingPreferencesDataSource
+import com.gdavidpb.tuindice.utils.AuthorizationInterceptor
+import com.gdavidpb.tuindice.data.source.settings.PreferencesDataSource
 import com.gdavidpb.tuindice.domain.usecase.GetUpdateInfoUseCase
 import com.gdavidpb.tuindice.domain.usecase.RequestReviewUseCase
 import com.gdavidpb.tuindice.domain.usecase.SetLastScreenUseCase
